CTA Poem
This is my call to action peom from my senior year of highschool. It refers to the roads in Chuuk which are very bumpy and aren't paved because of the misuse of infrastructure money in the government by corrupted leaders. Barely anyone wants to be magis and fix the roads because landowners are very greedy. This is a poem to draw attention to this problem which is overlooked since it has become a norm, but as a student who had to drive through the rough roads to get to my school which resides on a hill, it was very tiring. This is a poem to call for action; cooperation, teamwork, compassion, determination, and agreement.